Develop Log Test Code and modify config Class

+++ src/main/java/science/workbook/repository/repositoryMongo/LogRepository.java
@@ -3,5 +3,7 @@ package science.workbook.repository.repositoryMongo;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import science.workbook.domain.Log;
 
-public interface LogRepository extends MongoRepository<Log, Long> {
+import java.math.BigInteger;
+
+public interface LogRepository extends MongoRepository<Log, BigInteger> {
 }
